added info messages + lazy downloading to download.py

+++ src/easycoref/download.py
@@ -1,3 +1,5 @@
+from genericpath import exists
+import pathlib
 import subprocess
 from sys import platform
 from urllib.request import urlretrieve
@@ -14,29 +16,53 @@ def download():
     # Create a hidden directory in HOME
     home_path = Path.home()
     easycoref_path = os.path.join(home_path, '.easycoref')
-    os.makedirs(easycoref_path, mode=511, exist_ok=True)
+    if os.path.exists(easycoref_path):
+        logging.info('Ressources already cached, at least in part')
+    else:
+        logging.info('Ressources not downloaded yet, starting downloads...')
+        os.makedirs(easycoref_path, mode=511)
 
     # Download pretrained embeddings.
-    url = 'http://downloads.cs.stanford.edu/nlp/data/glove.840B.300d.zip'
-    dst = os.path.join(easycoref_path, 'glove')  # TODO : change embeddings path in e2ecoref
-    os.makedirs(dst, mode=511, exist_ok=True)
-    local_filename, headers = urlretrieve(url)
-    with zipfile.ZipFile(local_filename,"r") as zip_ref:
-        zip_ref.extractall(dst)
+    dst = os.path.join(easycoref_path, 'glove')
+    if os.path.exists(os.path.join(dst, 'glove.840B.300d.txt')):
+        logging.info('1/3 : Glove Embeddings already downloaded!')
+
+    else:
+        logging.info('1/3 : Starting Glove Embeddings Download...')
+        url = 'http://downloads.cs.stanford.edu/nlp/data/glove.840B.300d.zip'
+        os.makedirs(dst, mode=511, exist_ok=True)
+        local_filename, headers = urlretrieve(url)
+        with zipfile.ZipFile(local_filename,"r") as zip_ref:
+            zip_ref.extractall(dst)
+        logging.info('Done downloading!')
 
     # Download char vocab for e2ecoref
-    url = 'https://lil.cs.washington.edu/coref/char_vocab.english.txt'
     dst = os.path.join(easycoref_path, 'char_vocab.english.txt') 
-    urlretrieve(url, dst)
+    if os.path.exists(dst):
+        logging.info('2/3 : Char Vocab already downloaded!')
+    else:
+        logging.info('2/3 : Starting Char Vocab Download...')
+        url = 'https://lil.cs.washington.edu/coref/char_vocab.english.txt'
+        urlretrieve(url, dst)
+        logging.info('Done downloading!')
 
-    
-    url_e2elogs = "https://docs.google.com/uc?export=download&id=1fkifqZzdzsOEo0DXMzCFjiNXqsKG_cHi"
-    dst = os.path.join(easycoref_path, 'e2e_logs')  # TODO : change embeddings path in e2ecoref
-    os.makedirs(dst, mode=511, exist_ok=True)
-    local_filename, headers = urlretrieve(url_e2elogs)
-    with tarfile.TarFile(local_filename,"r:gz") as tar_ref:
-        tar_ref.extractall(dst)
+    dst = os.path.join(easycoref_path, 'e2e_logs')
+    if not(os.path.exists(dst)):
+        os.makedirs(dst, mode=511)
 
+    if any(Path(dst).iterdir()):
+        logging.info('3/3 : e2e-coref logs already downloaded!')
+    else:
+        logging.info('3/3 : Starting e2e-coref logs Download...')
+        url_e2elogs = "https://docs.google.com/uc?export=download&id=1fkifqZzdzsOEo0DXMzCFjiNXqsKG_cHi"
+        local_filename, headers = urlretrieve(url_e2elogs)
+        logging.info('Done downloading!')
+        logging.info('Unzipping...')
+        with tarfile.TarFile(local_filename,"r:gz") as tar_ref:
+            tar_ref.extractall(dst)
+            logging.info('Unzipped!')
+
+        logging.info('All done!')
 
     # Compile tensorflow custom C ops 
     subprocess.check_call("TF_CFLAGS=( $(python -c 'import tensorflow as tf; print(" ".join(tf.sysconfig.get_compile_flags()))') )".split())
